semester_7/ai/cursach/requirements.txt
dostoevsky
python-telegram-bot